requirements.txt
newsapi-python==0.2.7
python-dotenv==1.0.1
pandas==2.2.3
numpy==2.1.3
nltk==3.9.1
scikit-learn==1.5.2
textblob==0.18.0.post0
matplotlib==3.9.2
seaborn==0.13.2
requests==2.32.3
newspaper3k==0.2.8
lxml[html_clean]==5.1.0
wordcloud==1.9.3
pyLDAvis==3.4.1
funcy==1.18
numexpr>=2.7.3